+++ packages/ndla-ui/src/Footer/Footer.tsx
@@ -8,115 +8,15 @@
 
 import React from 'react';
 import styled from '@emotion/styled';
-import { colors, spacing, fonts, misc, mq, breakpoints } from '@ndla/core';
+import { colors, spacing, fonts, mq, breakpoints } from '@ndla/core';
 // @ts-ignore
 import { injectT } from '@ndla/i18n';
 // @ts-ignore
 import { FooterHeaderIcon } from '@ndla/icons/common';
 // @ts-ignore
-import Modal, { ModalHeader, ModalBody, ModalCloseButton } from '@ndla/modal';
-// @ts-ignore
 import { OneColumn } from '../Layout';
-
-import PrivacyNb from './privacy_nb';
-import PrivacyNn from './privacy_nn';
-import PrivacyEn from './privacy_en';
 import FooterLinks from './FooterLinks';
-
-const StyledFooterText = styled.p`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  > span {
-    padding: ${spacing.xsmall} 0;
-    text-align: center;
-  }
-  ${mq.range({ until: breakpoints.tabletWide })} {
-    ${fonts.sizes(16, 1.5)};
-    > span {
-      padding: 0;
-    }
-  }
-  ${mq.range({ until: breakpoints.mobileWide })} {
-    ${fonts.sizes(14, 1.3)};
-    > span {
-      padding-bottom: ${spacing.xsmall};
-    }
-  }
-`;
-
-type FooterTextProps = {
-  children: React.ReactNode;
-};
-
-export const FooterText: React.FunctionComponent<FooterTextProps> = ({
-  children,
-}) => <StyledFooterText>{children}</StyledFooterText>;
-
-type EditorNameProps = {
-  title: string;
-  name: string;
-};
-
-export const EditorName: React.FunctionComponent<EditorNameProps> = ({
-  title,
-  name,
-}) => (
-  <span>
-    <strong>{title}</strong> {name}
-  </span>
-);
-
-type FooterPrivacyProps = {
-  label: string;
-  lang: 'nb' | 'nn' | 'en';
-};
-
-const privacyTexts = (lang: string) => {
-  if (lang === 'nn') {
-    return <PrivacyNn />;
-  }
-  if (lang === 'en') {
-    return <PrivacyEn />;
-  }
-  return <PrivacyNb />;
-};
-
-const StyledPrivacyButton = styled.button`
-  background: none;
-  color: #fff;
-  border: 0;
-  padding: 0;
-  box-shadow: ${misc.textLinkBoxShadow};
-  cursor: pointer;
-  margin-bottom: ${spacing.large};
-  &:hover,
-  &:focus {
-    box-shadow: none;
-  }
-`;
-
-const FooterPrivacy: React.FunctionComponent<FooterPrivacyProps> = ({
-  lang,
-  label,
-}) => (
-  <StyledFooterText>
-    <Modal
-      activateButton={
-        <StyledPrivacyButton type="button">{label}</StyledPrivacyButton>
-      }
-      size="fullscreen">
-      {(onClose: void) => (
-        <OneColumn cssModifier="medium">
-          <ModalHeader>
-            <ModalCloseButton onClick={onClose} title="Lukk" />
-          </ModalHeader>
-          <ModalBody>{privacyTexts(lang)}</ModalBody>
-        </OneColumn>
-      )}
-    </Modal>
-  </StyledFooterText>
-);
+import FooterPrivacy from './FooterPrivacy';
 
 const StyledFooter = styled.footer`
   background: ${colors.brand.dark};
